+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [markdownlint](https://dlaa.me/markdownlint/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.7] - 2023-06-07
+
+### Added in 1.6.7
+
+- Improved detection of license file or string and use in container(s)
+- Warning message if SENZING_ENGINE_CONFIGURATION_JSON is set
+
 ## [1.6.6] - 2023-02-06
 
 ### Changed in 1.6.6